backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-multipart==0.0.6
bcrypt==4.0.1
PyJWT==2.8.0
python-decouple==3.8
stripe==7.8.0  # Dodane